+++ lib/api.dart
@@ -2,9 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library rpc;
+library rpc.api;
 
 export 'src/annotations.dart';
+export 'src/context.dart';
 export 'src/errors.dart';
-export 'src/message.dart';
-export 'src/server.dart';