+++ lib/src/errors.dart
@@ -4,8 +4,6 @@
 
 library rpc.errors;
 
-import 'dart:io';
-
 class RpcError implements Exception {
 
   final int statusCode;
@@ -19,23 +17,21 @@ class RpcError implements Exception {
 
 class NotFoundError extends RpcError {
   NotFoundError([String message = "Not found."])
-      : super(HttpStatus.NOT_FOUND, 'Not Found', message);
+      : super(404, 'Not Found', message);
 }
 
 class BadRequestError extends RpcError {
   BadRequestError([String message = "Bad request."])
-      : super(HttpStatus.BAD_REQUEST, 'Bad Request', message);
+      : super(400, 'Bad Request', message);
 }
 
 class InternalServerError extends RpcError {
   InternalServerError([String message = "Internal Server Error."])
-      : super(HttpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', message);
+      : super(500, 'Internal Server Error', message);
 }
 
 class ApplicationError extends RpcError {
   ApplicationError(Exception e)
-      : super(HttpStatus.INTERNAL_SERVER_ERROR,
-              'Application Invocation Error',
-              e.toString());
+      : super(500,  'Application Invocation Error', e.toString());
 
 }